FIX: model int -> float

+++ Model.hpp
@@ -23,12 +23,12 @@ class Model {
   int nverts();
   int nfaces();
   Vec3f normal(int iface, int nthvert);
-  Vec3f normal(Vec2i uv);
+  Vec3f normal(Vec2f uv);
   Vec3f vert(int i);
   Vec3f vert(int iface, int nthvert);
-  Vec2i uv(int iface, int nthvert);
-  TGAColor diffuse(Vec2i uv);
-  float specular(Vec2i uv);
+  Vec2f uv(int iface, int nthvert);
+  TGAColor diffuse(Vec2f uv);
+  float specular(Vec2f uv);
   std::vector<int> face(int idx);
 };
